+++ cmd/root.go
@@ -138,7 +138,7 @@ func init() {
 	rootCmd.PersistentFlags().StringVar(&defaults.CustomConfigurationFile, globals.ConfigLabel, defaults.ConfigurationFile, "configuration file")
 	setPflag(rootCmd, globals.SandboxHomeLabel, "", "SANDBOX_HOME", defaults.Defaults().SandboxHome, "Sandbox deployment directory", false)
 	setPflag(rootCmd, globals.SandboxBinaryLabel, "", "SANDBOX_BINARY", defaults.Defaults().SandboxBinary, "Binary repository", false)
-	setPflag(rootCmd, globals.ShellPathLabel, "", "SHELL_PATH", globals.ShellPathValue, "Which shell to use for generated scripts", false)
+	setPflag(rootCmd, globals.ShellPathLabel, "", "SHELL_PATH", common.Which("bash"), "Which shell to use for generated scripts", false)
 	rootCmd.PersistentFlags().BoolP(globals.SkipLibraryCheck, "", false, "Skip check for needed libraries (may cause nasty errors)")
 
 	rootCmd.InitDefaultVersionFlag()